+++ config/turtle.php
@@ -7,6 +7,47 @@ return [
         'frontend' => true,
         'registration' => true,
         'contact' => true,
+        'billing' => true,
+    ],
+
+    // billing
+    'billing' => [
+        'stripe_secret_key' => 'sk_test_QSlbZsEyVA0pRaqyWKduKRYT',
+        'trial_period' => '30 days',
+        'plans' => [
+            // Stripe plan ID => [options]
+            'basic' => [
+                'name' => 'Basic',
+                'description' => 'A great starter plan for individuals.',
+                'price' => '$10',
+                'period' => 'month',
+                'limits' => [
+                    // model => limit
+                    'App\Post' => 5,
+                    'App\Page' => 5,
+                ],
+            ],
+            'plus' => [
+                'name' => 'Plus',
+                'description' => 'A bit more flexible, ideal for teams.',
+                'price' => '$20',
+                'period' => 'month',
+                'limits' => [
+                    'App\Post' => 25,
+                    'App\Page' => 25,
+                ],
+            ],
+            'premium' => [
+                'name' => 'Premium',
+                'description' => 'Fully unlimited! Perfect for companies.',
+                'price' => '$50',
+                'period' => 'month',
+                'limits' => [
+                    'App\Post' => null, // null = unlimited
+                    'App\Page' => null, // null = unlimited
+                ],
+            ],
+        ],
     ],
 
     // demo mode
@@ -27,6 +68,7 @@ return [
     ],
     'models' => [
         'activity' => 'Kjdion84\Turtle\Models\Activity',
+        'billing' => 'Kjdion84\Turtle\Models\Billing',
         'permission' => 'Kjdion84\Turtle\Models\Permission',
         'role' => 'Kjdion84\Turtle\Models\Role',
         'user' => 'App\User',