+++ resources/views/layouts/app.blade.php
@@ -17,7 +17,7 @@
 <body{!! session('flash') ? ' data-flash-class="'.session('flash.0').'" data-flash-message="'.session('flash.1').'"' : '' !!}>
 <nav class="navbar navbar-expand-lg navbar-dark bg-primary">
     <div class="container">
-        <a class="navbar-brand" href="{{ route('index') }}">{{ config('app.name') }}</a>
+        <a class="navbar-brand" href="{{ route('index') }}"><i class="fa fa-cog"></i> {{ config('app.name') }}</a>
         <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbar">
             <span class="navbar-toggler-icon"></span>
         </button>
@@ -52,6 +52,9 @@
                     <li class="nav-item dropdown">
                         <a class="nav-link dropdown-toggle" href="#" data-toggle="dropdown"><i class="fa fa-user-circle"></i> {{ auth()->user()->name }}</a>
                         <div class="dropdown-menu dropdown-menu-right">
+                            @if(auth()->user()->billable && config('turtle.allow.billing'))
+                                <a class="dropdown-item{{ request()->is('billing') ?  ' active' : '' }}" href="{{ route('billing') }}">Billing</a>
+                            @endif
                             <a class="dropdown-item{{ request()->is('profile') ?  ' active' : '' }}" href="{{ route('profile') }}">Edit Profile</a>
                             <a class="dropdown-item{{ request()->is('password/change') ?  ' active' : '' }}" href="{{ route('password.change') }}">Change Password</a>
                             <a class="dropdown-item" href="{{ route('logout') }}">Logout</a>
@@ -63,6 +66,7 @@
     </div>
 </nav>
 
+@include('turtle::auth.billing.alert')
 @yield('content')
 
 <footer>