+++ resources/views/users/add.blade.php
@@ -35,6 +35,16 @@
                 </select>
             </div>
 
+            <input type="hidden" name="billable" value="0">
+            @if(config('turtle.allow.billing'))
+                <div class="form-check">
+                    <label class="form-check-label">
+                        <input type="checkbox" name="billable" id="billable" class="form-check-input" value="1">
+                        Billable
+                    </label>
+                </div>
+            @endif
+
             <div class="form-group">
                 <div>
                     <label>Roles</label>