+++ resources/views/users/edit.blade.php
@@ -26,6 +26,16 @@
                 </select>
             </div>
 
+            <input type="hidden" name="billable" value="0">
+            @if(config('turtle.allow.billing'))
+                <div class="form-check">
+                    <label class="form-check-label">
+                        <input type="checkbox" name="billable" id="billable" class="form-check-input" value="1"{{ $user->billable ? ' checked' : '' }}>
+                        Billable
+                    </label>
+                </div>
+            @endif
+
             <div class="form-group">
                 <div>
                     <label>Roles</label>